backend/requirements.txt
fastapi
uvicorn
yfinance
langchain-groq
langchain-core
langgraph
tavily-python
python-dotenv
pydantic